Organizing the gem, class names and files
* IMHO the document class should have your own place inside of the
gem since it is a gem convention
